refactor(#80): 로비 Redis 조회 책임 분리
